requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-amazon==9.22.0
apache-airflow-providers-common-compat==1.14.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.32.0
apache-airflow-providers-http==6.0.0
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.12.0
apache-airflow-task-sdk==1.1.7
argcomplete==3.6.3
arro3-core==0.8.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==25.4.0
aws-embedded-metrics==3.3.1
awswrangler==3.15.1
babel==2.18.0
beautifulsoup4==4.14.3
boto3==1.42.60
botocore==1.42.60
cachetools==6.2.6
cadwyn==5.4.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.5
deltalake==1.4.2
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
duckdb==1.4.4
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.24
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenback==1.3.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.2.1
more-itertools==10.8.0
msgspec==0.20.0
multidict==6.7.1
natsort==8.4.0
numpy==2.2.6
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pendulum==3.2.0
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAthena==3.30.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydynamodb==0.8.1
Pygments==2.19.2
pygtrie==2.5.0
pyiceberg==0.11.1
PyJWT==2.11.0
pymssql==2.3.13
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyroaring==1.0.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
redshift-connector==2.1.12
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
sagemaker_studio==1.1.8
scramp==1.4.8
setproctitle==1.3.7
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.3.0
snowflake-sqlalchemy==1.9.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlalchemy-bigquery==1.16.0
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
strictyaml==1.7.3
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
tomlkit==0.14.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
watchtower==3.4.0
websockets==16.0
wirerope==1.0.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
